pyproject.toml
[project]
name = "wishket"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gpustat>=1.1.1",
    "numpy>=2.4.2",
    "opencv-python>=4.13.0.90",
    "optuna>=4.7.0",
    "psutil>=7.2.2",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "torch>=2.10.0",
    "ultralytics>=8.4.9",
    "wandb>=0.24.1",
]
